+++ CHANGELOG.md
@@ -23,8 +23,9 @@ Toast.open('Toasty!')
 
 * **New component: Slider**, [see docs](https://buefy.org/documentation/slider) (thanks @yxngl)
 * **New component: Navbar**, [see docs](https://buefy.org/documentation/navbar) (thanks @wanxe)
-* **New component: Progress Bar**, [see docs](https://buefy.org/documentation/progress)
+* **New component: Progress Bar**, [see docs](https://buefy.org/documentation/progress) (thanks @service-paradis)
 * Improve to tree shaking when importing individual components
+* Add ``header`` slot to table column
 * Add ``defaultTooltipDelay`` constructor option (thanks @service-paradis)
 * Add ``sort-icon`` and ``sort-icon-size`` props to table (@modularcoder)
 * Add ``default-minutes`` and ``default-seconds`` to timepicker